Move getVerdicts to ConfigurationService
Also add some documentation.
